Make REST calls by DELETE idempotent

+++ main/java/sportbets/service/authorization/impl/CommunityRoleServiceImpl.java
@@ -89,6 +89,9 @@ public class CommunityRoleServiceImpl implements CommunityRoleService {
     @Override
     @Transactional
     public void deleteById(Long id) {
-        roleRepository.deleteById(id);
+        // Idempotent check: If it is already gone, do nothing
+        if (roleRepository.existsById(id)) {
+            roleRepository.deleteById(id);
+        }
     }
 }
